Calculate entity lookups once
