+++ src/components/LocusSelection.js
@@ -9,63 +9,67 @@ const LocusSelection = ({
   selectedTagVariants,
   selectedIndexVariants,
   selectedStudies,
+  lookups,
   handleDeleteGene,
   handleDeleteTagVariant,
   handleDeleteIndexVariant,
   handleDeleteStudy,
-}) => (
-  <div>
-    <Text>
-      {selectedGenes
-        ? selectedGenes.map(d => (
-            <Chip
-              key={d}
-              label={d}
-              type="gene"
-              onDelete={handleDeleteGene(d)}
-            />
-          ))
-        : null}
-      {/* {selectedGenes &&
+}) => {
+  const { geneDict, studyDict } = lookups;
+  return (
+    <div>
+      <Text>
+        {selectedGenes
+          ? selectedGenes.map(d => (
+              <Chip
+                key={d}
+                label={geneDict[d].symbol}
+                type="gene"
+                onDelete={handleDeleteGene(d)}
+              />
+            ))
+          : null}
+        {/* {selectedGenes &&
       (selectedTagVariants || selectedIndexVariants || selectedStudies)
         ? AND
         : null} */}
-      {selectedTagVariants
-        ? selectedTagVariants.map(d => (
-            <Chip
-              key={d}
-              label={d}
-              type="tagVariant"
-              onDelete={handleDeleteTagVariant(d)}
-            />
-          ))
-        : null}
-      {/* {selectedTagVariants && (selectedIndexVariants || selectedStudies)
+        {selectedTagVariants
+          ? selectedTagVariants.map(d => (
+              <Chip
+                key={d}
+                label={d}
+                type="tagVariant"
+                onDelete={handleDeleteTagVariant(d)}
+              />
+            ))
+          : null}
+        {/* {selectedTagVariants && (selectedIndexVariants || selectedStudies)
         ? AND
         : null} */}
-      {selectedIndexVariants
-        ? selectedIndexVariants.map(d => (
-            <Chip
-              key={d}
-              label={d}
-              type="indexVariant"
-              onDelete={handleDeleteIndexVariant(d)}
-            />
-          ))
-        : null}
-      {/* {selectedIndexVariants && selectedStudies ? AND : null} */}
-      {selectedStudies
-        ? selectedStudies.map(d => (
-            <Chip
-              key={d}
-              label={d}
-              type="study"
-              onDelete={handleDeleteStudy(d)}
-            />
-          ))
-        : null}
-    </Text>
-  </div>
-);
+        {selectedIndexVariants
+          ? selectedIndexVariants.map(d => (
+              <Chip
+                key={d}
+                label={d}
+                type="indexVariant"
+                onDelete={handleDeleteIndexVariant(d)}
+              />
+            ))
+          : null}
+        {/* {selectedIndexVariants && selectedStudies ? AND : null} */}
+        {selectedStudies
+          ? selectedStudies.map(d => (
+              <Chip
+                key={d}
+                label={studyDict[d].traitReported}
+                type="study"
+                onDelete={handleDeleteStudy(d)}
+              />
+            ))
+          : null}
+      </Text>
+    </div>
+  );
+};
 
 export default LocusSelection;
